로그인에 따른 권한 수정

+++ Board/src/main/java/com/study/domain/post/PostController.java
@@ -9,9 +9,11 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.SessionAttribute;
 
 import com.study.common.dto.MessageDto;
 import com.study.common.dto.SearchDto;
+import com.study.domain.user.UserResponse;
 import com.study.paging.PagingResponse;
 
 import lombok.RequiredArgsConstructor;
@@ -24,8 +26,12 @@ public class PostController {
     
     // 정보 게시글 작성 페이지
     @GetMapping("/post/write.do")
-    public String openInfoPostWrite(@RequestParam(value = "id", required = false) final Long id, Model model) {
-        if (id != null) {
+    public String openInfoPostWrite(@RequestParam(value = "id", required = false) final Long id,
+    		@SessionAttribute(name = "userInfo", required = false)UserResponse user, Model model) {
+    	if(user == null) {
+        	return "user/needLogin";
+        }
+    	if (id != null) {
             PostResponse post = postService.findById(id);
             model.addAttribute("post", post);
         }
@@ -35,9 +41,13 @@ public class PostController {
         
     // 신규 게시글 생성
     @PostMapping("/post/save.do")
-    public String savePost(final PostRequest params, HttpSession session, Model model) {
+    public String savePost(final PostRequest params,
+    		@SessionAttribute(name = "userInfo", required = false)UserResponse user, HttpSession session, Model model) {
 //    	int writer = (int) session.getAttribute("no");
-    	params.setWriterNo(1);
+    	if(user == null) {
+        	return "user/needLogin";
+        }
+    	params.setWriterNo(user.getUserNo());
         postService.savePost(params);
         if (params.getType() == 0) {
         	 MessageDto message = new MessageDto("게시글 생성이 완료되었습니다.", "/post/infolist.do", RequestMethod.GET, null);
@@ -66,16 +76,25 @@ public class PostController {
     
     // 게시글 상세 페이지
     @GetMapping("/post/view.do")
-    public String openPostView(@RequestParam final Long id, Model model) {
+    public String openPostView(@RequestParam final Long id
+    		,@SessionAttribute(name = "userInfo", required = false)UserResponse user, Model model) {
+    	if(user == null) {
+        	return "user/needLogin";
+        }
         PostResponse post = postService.findById(id);
         model.addAttribute("post", post);
+        model.addAttribute("userInfo",user);
         return "post/view";
     }
     
     // 기존 게시글 수정
     @PostMapping("/post/update.do")
-    public String updatePost(final PostRequest params, Model model) {
-    	params.setWriterNo(1);
+    public String updatePost(final PostRequest params
+    		,@SessionAttribute(name = "userInfo", required = false)UserResponse user, Model model) {
+    	if(user == null) {
+        	return "user/needLogin";
+        }
+    	params.setWriterNo(user.getUserNo());
         postService.updatePost(params);
         
         if (params.getType() == 0) {
@@ -89,8 +108,12 @@ public class PostController {
     
     // 게시글 삭제
     @PostMapping("/post/delete.do")
-    public String deletePost(@RequestParam final Long id, @RequestParam final int type, Model model) {
-        postService.deletePost(id);
+    public String deletePost(@RequestParam final Long id
+    		,@SessionAttribute(name = "userInfo", required = false)UserResponse user, @RequestParam final int type, Model model) {
+    	if(user == null) {
+        	return "user/needLogin";
+        }
+    	postService.deletePost(id);
         if(type == 0) {
         	MessageDto message = new MessageDto("게시글 삭제가 완료되었습니다.", "/post/infolist.do", RequestMethod.GET, null);
         	return showMessageAndRedirect(message, model);
